backend/requirements.txt
absl-py==2.1.0
asttokens==2.4.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
decorator==4.4.2
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.13.1
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.55.0
fsspec==2024.2.0
grpcio==1.68.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
kiwisolver==1.4.7
llvmlite==0.43.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
openai-whisper==20240930
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
proglog==0.1.10
prompt_toolkit==3.0.48
protobuf==4.21.2
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tiktoken==0.8.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.0
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
